+++ src/direct_goal.cpp
@@ -2,7 +2,25 @@
 // Created by hyx020222 on 7/26/24.
 //
 
-#include "direct_goal.h"
+#include <ros/ros.h>
+#include <Eigen/Eigen>
+
+#include <message_filters/subscriber.h>
+#include <message_filters/sync_policies/approximate_time.h>
+#include <message_filters/sync_policies/exact_time.h>
+#include <message_filters/time_synchronizer.h>
+
+#include <geometry_msgs/PoseStamped.h>
+#include <nav_msgs/Odometry.h>
+
+#include <easondrone_msgs/ControlCommand.h>
+
+
+easondrone_msgs::ControlCommand easondrone_ctrl_cmd_;                      //无人机当前执行命令
+
+ros::Subscriber goal_sub_, odom_sub_;
+ros::Publisher easondrone_ctrl_cmd_pub_;
+
 
 void goal_cb(const geometry_msgs::PoseStamped::ConstPtr& msg){
     easondrone_ctrl_cmd_.Reference_State.position_ref[0] = msg->pose.position.x;
@@ -17,6 +35,7 @@ void goal_cb(const geometry_msgs::PoseStamped::ConstPtr& msg){
 void odometryCallback(const nav_msgs::OdometryConstPtr &msg){
     easondrone_ctrl_cmd_.Reference_State.position_ref[2] = msg->pose.pose.position.z;
 }
+
 
 int main(int argc, char **argv) {
     ros::init(argc, argv, "direct_goal");